Extract out responsibility of formatting text from TextUI class to Formatter class
