+++ src/seedu/addressbook/ui/Formatter.java
@@ -1,0 +1,52 @@
+package seedu.addressbook.ui;
+
+import java.util.List;
+
+/**
+ * Format text for display.
+ */
+public class Formatter {
+
+    /** A decorative prefix added to the beginning of lines printed by AddressBook */
+    public static final String LINE_PREFIX = "|| ";
+
+    /** A platform independent line separator. */
+    public static final String LS = System.lineSeparator();
+
+    public static final String DIVIDER = "===================================================";
+
+    /** Format of indexed list item */
+    public static final String MESSAGE_INDEXED_LIST_ITEM = "\t%1$d. %2$s";
+
+
+    /** Offset required to convert between 1-indexing and 0-indexing.  */
+    public static final int DISPLAYED_INDEX_OFFSET = 1;
+
+    /** Format of a comment input line. Comment lines are silently consumed when reading user input. */
+    public static final String COMMENT_LINE_FORMAT_REGEX = "#.*";
+
+    /** Shows a list of strings to the user, formatted as an indexed list. */
+    public static String showToUserAsIndexedList(List<String> list) {
+        return getIndexedListForViewing(list);
+    }
+
+    /** Formats a list of strings as a viewable indexed list. */
+    private static String getIndexedListForViewing(List<String> listItems) {
+        final StringBuilder formatted = new StringBuilder();
+        int displayIndex = 0 + DISPLAYED_INDEX_OFFSET;
+        for (String listItem : listItems) {
+            formatted.append(getIndexedListItem(displayIndex, listItem)).append("\n");
+            displayIndex++;
+        }
+        return formatted.toString();
+    }
+
+    /**
+     * Formats a string as a viewable indexed list item.
+     *
+     * @param visibleIndex visible index for this listing
+     */
+    private static String getIndexedListItem(int visibleIndex, String listItem) {
+        return String.format(MESSAGE_INDEXED_LIST_ITEM, visibleIndex, listItem);
+    }
+}
